Changes necessary to correspond with the changes made in Amod to work with the shortest path library. Improves accuracy of the travel times obtained by the existing providers.

+++ src/main/java/cz/cvut/fel/aic/agentpolis/simmodel/entity/vehicle/PhysicalVehicle.java
@@ -63,8 +63,8 @@ public class PhysicalVehicle extends Vehicle {
 	}
 
 	@Override
-	public int getVelocity() {
-		return maxVelocity;
+	public double getVelocity() {
+		return (double) maxVelocity;
 	}
 
 }
